dedup flatten
