+++ src/rpc/client.rs
@@ -1,3 +1,20 @@
 //! RPC Client for docs and authors
+use anyhow::Result;
+use futures_util::{Stream, StreamExt};
+
 pub mod authors;
 pub mod docs;
+
+fn flatten<T, E1, E2>(
+    s: impl Stream<Item = Result<Result<T, E1>, E2>>,
+) -> impl Stream<Item = Result<T>>
+where
+    E1: std::error::Error + Send + Sync + 'static,
+    E2: std::error::Error + Send + Sync + 'static,
+{
+    s.map(|res| match res {
+        Ok(Ok(res)) => Ok(res),
+        Ok(Err(err)) => Err(err.into()),
+        Err(err) => Err(err.into()),
+    })
+}
