+++ src/rpc/client/docs.rs
@@ -35,6 +35,8 @@ use crate::{
     RecordIdentifier,
 };
 
+use super::flatten;
+
 /// Iroh docs client.
 #[derive(Debug, Clone)]
 pub struct Client<C = BoxedConnector<RpcService>> {
@@ -731,20 +733,6 @@ impl Stream for ExportFileProgress {
     fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
         Pin::new(&mut self.stream).poll_next(cx)
     }
-}
-
-fn flatten<T, E1, E2>(
-    s: impl Stream<Item = Result<Result<T, E1>, E2>>,
-) -> impl Stream<Item = Result<T>>
-where
-    E1: std::error::Error + Send + Sync + 'static,
-    E2: std::error::Error + Send + Sync + 'static,
-{
-    s.map(|res| match res {
-        Ok(Ok(res)) => Ok(res),
-        Ok(Err(err)) => Err(err.into()),
-        Err(err) => Err(err.into()),
-    })
 }
 
 #[cfg(test)]
